requirements.txt
python-telegram-bot==21.6
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.0
